changed source to python 3

+++ wideresnet.py
@@ -28,8 +28,9 @@ class BasicBlock(nn.Module):
         self.droprate = dropRate
         self.equalInOut = (in_planes == out_planes)
         self.convShortcut = (not self.equalInOut) and nn.Conv2d(in_planes, out_planes, kernel_size=1, stride=stride,
-                               padding=0, bias=False) or None
-        self.activateBeforeResidual= actBeforeRes
+                                                                padding=0, bias=False) or None
+        self.activateBeforeResidual = actBeforeRes
+
     def forward(self, x):
         if not self.equalInOut and self.activateBeforeResidual:
             x = self.relu1(self.bn1(x))
@@ -44,34 +45,43 @@ class BasicBlock(nn.Module):
         res = self.convShortcut(x) if not self.equalInOut else x
         return torch.add(res, out)
 
+
 class NetworkBlock(nn.Module):
     def __init__(self, nb_layers, in_planes, out_planes, block, stride, dropRate=0.0, leakyRate=0.01, actBeforeRes=True):
         super(NetworkBlock, self).__init__()
-        self.layer = self._make_layer(block, in_planes, out_planes, nb_layers, stride, dropRate, leakyRate, actBeforeRes)
+        self.layer = self._make_layer(
+            block, in_planes, out_planes, nb_layers, stride, dropRate, leakyRate, actBeforeRes)
+
     def _make_layer(self, block, in_planes, out_planes, nb_layers, stride, dropRate, leakyRate, actBeforeRes):
         layers = []
         for i in range(nb_layers):
-            layers.append(block(i == 0 and in_planes or out_planes, out_planes, i == 0 and stride or 1, dropRate, leakyRate, actBeforeRes))
+            layers.append(block(i == 0 and in_planes or out_planes, out_planes,
+                                i == 0 and stride or 1, dropRate, leakyRate, actBeforeRes))
         return nn.Sequential(*layers)
+
     def forward(self, x):
         return self.layer(x)
+
 
 class WideResNet(nn.Module):
     def __init__(self, depth, num_classes, widen_factor=1, dropRate=0.0, leakyRate=0.01):
         super(WideResNet, self).__init__()
         nChannels = [16, 16*widen_factor, 32*widen_factor, 64*widen_factor]
         assert((depth - 4) % 6 == 0)
-        n = (depth - 4) / 6
+        n = int((depth - 4) / 6)
         block = BasicBlock
         # 1st conv before any network block
         self.conv1 = nn.Conv2d(3, nChannels[0], kernel_size=3, stride=1,
                                padding=1, bias=False)
         # 1st block
-        self.block1 = NetworkBlock(n, nChannels[0], nChannels[1], block, 1, dropRate, leakyRate, False)
+        self.block1 = NetworkBlock(
+            n, nChannels[0], nChannels[1], block, 1, dropRate, leakyRate, False)
         # 2nd block
-        self.block2 = NetworkBlock(n, nChannels[1], nChannels[2], block, 2, dropRate, leakyRate, True)
+        self.block2 = NetworkBlock(
+            n, nChannels[1], nChannels[2], block, 2, dropRate, leakyRate, True)
         # 3rd block
-        self.block3 = NetworkBlock(n, nChannels[2], nChannels[3], block, 2, dropRate, leakyRate, True)
+        self.block3 = NetworkBlock(
+            n, nChannels[2], nChannels[3], block, 2, dropRate, leakyRate, True)
         # global average pooling and classifier
         self.bn1 = nn.BatchNorm2d(nChannels[3])
         self.relu = nn.LeakyReLU(leakyRate, inplace=True)
@@ -87,6 +97,7 @@ class WideResNet(nn.Module):
                 m.bias.data.zero_()
             elif isinstance(m, nn.Linear):
                 m.bias.data.zero_()
+
     def forward(self, x):
         out = self.conv1(x)
         out = self.block1(out)
